Import led_ex1_blinky as base for c2000-sine
This is an initial Code Composer Studio 10.2.0.00009 project
for the C2000 port of stm32-sine.

The base project comes from the driverlib/led_ex1_blinky
example in C2000Ware 3.03.00.00.

Add the define _LAUNCHXL_F28379D to RAM and FLASH
targets to ensure the correct frequency is used on the
LAUNCHXL-F28379D dev board.

Tests:
- Build and run project on LAUNCHXL-F28379D