Fix the definition of NULL for C++ code
The definition of NULL has to be just 0. It can't be typed as a
void* as it is in C. This changes the definition to follow that.

The include paths are changed to remove incorrect out of tree
references to C2000Ware. The devicelib code used by this
project is contained within the project (as it is 3-clause BSD
licensed this is OK for a GPL v3 licensed project).

Tests:
 - Complete clean build with definition of TermCmds in
   terminal_prj restored to the stm32-sine import.