Update the terminal support to support the C2000 SCI
This adds terminal support by re-introducing the hwinit
module to set up the C2000 SCI/UART on the LAUNCHXL-
F28379D. The terminal.cpp driver is rewritten to support
polled FIFO UART support rather than be DMA driven like the
STM32 (it doesn't appear to be possible to drive the F28379D
UART via DMA). The old blinky main loop in c2000-sine is
replaced with a call to term_Run() as per the stm32-sine
main().

Tests:
 - Run the project from RAM testing various valid commands
 - Test invalid commands and blank lines
 - Check characters types are echoed back